BDOG-523 Refactor cache flavours
